Chore: 데이터를 model 폴더에서 asset 폴더로 이동, 연관된 파일 수정
