+++ src/app/api/membership-information/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 
-import { membershipInformation } from '@/model/membershipInformation.ts';
+import { membershipInformation } from '@/assets/membershipInformation.ts';
 
 export function GET() {
   return NextResponse.json(membershipInformation);
