--- src/assets/CommonBenefits.ts
@@ -1,20 +1,0 @@
-export const commonBenefitsData = [
-  {
-    id: '1',
-    title: '혜택 1',
-    benefit: '경조사비 2배 + α',
-    image: '',
-  },
-  {
-    id: '2',
-    title: '혜택 2',
-    benefit: '사이버 잡일 시키기',
-    image: '',
-  },
-  {
-    id: '3',
-    title: '혜택 3',
-    benefit: 'SNS 답장 빠른 반응',
-    image: '',
-  },
-];
