--- src/assets/SponsorsReviewData.ts
@@ -1,22 +1,0 @@
-export const sponsorsReviewData = [
-  {
-    name: '근박',
-    review: '이길수있어요. 힘내봐요. 서렌치지마요.',
-    rating: 5,
-  },
-  {
-    name: '기박',
-    review: '돈이 아깝고 머리가 아파요. 0점이 없어서 1점 드릴게요.',
-    rating: 1,
-  },
-  {
-    name: 'TEST',
-    review: '돈이 아깝고 머리가 아파요. 0점이 없어서 1점 드릴게요.',
-    rating: 1,
-  },
-  {
-    name: '김현',
-    review: '코스피 나스닥 빝코 다 떨어지고 있어요. 살려주세요',
-    rating: 1,
-  },
-];
